Validate AstralRT against UE 5.7 containers

+++ plugins/unreal/AstralRT/Source/AstralRT/Private/AstralEmbedder.cpp
@@ -91,7 +91,7 @@ bool UAstralEmbedder::EnqueueUtf8Bytes(const TArray<uint8>& Utf8Bytes, int64& Ou
     Text.data = Utf8Bytes.GetData();
     Text.len = static_cast<uint32_t>(Utf8Bytes.Num());
 
-    uint64 Ticket = 0;
+    uint64_t Ticket = 0;
     const AstralErr Err = astral_embed_enqueue(static_cast<AstralHandle>(EmbedderHandle), Text, &Ticket);
     if (Err != ASTRAL_OK)
     {
@@ -119,18 +119,18 @@ bool UAstralEmbedder::EnqueueImage(const FAstralImageDesc& Image, int64& OutTick
     AstralImageDesc Native{};
     Native.size = sizeof(AstralImageDesc);
     Native.format = static_cast<AstralImageFormat>(Image.Format);
-    Native.width = Image.Width;
-    Native.height = Image.Height;
-    Native.row_stride = Image.RowStride;
-    Native.flags = Image.Flags;
+    Native.width = static_cast<uint32_t>(Image.Width);
+    Native.height = static_cast<uint32_t>(Image.Height);
+    Native.row_stride = static_cast<uint32_t>(Image.RowStride);
+    Native.flags = static_cast<uint32_t>(Image.Flags);
     Native.pixels.data = Image.Pixels.GetData();
     Native.pixels.len = static_cast<uint32_t>(Image.Pixels.Num());
     Native.gpu_device = Image.GpuDevice;
-    Native.gpu_route_flags = Image.GpuRouteFlags;
-    Native.gpu_device_mask = Image.GpuDeviceMask;
+    Native.gpu_route_flags = static_cast<uint32_t>(Image.GpuRouteFlags);
+    Native.gpu_device_mask = static_cast<uint64_t>(Image.GpuDeviceMask);
     Native.gpu_stream = reinterpret_cast<void*>(static_cast<uintptr_t>(Image.GpuStream));
 
-    uint64 Ticket = 0;
+    uint64_t Ticket = 0;
     const AstralErr Err = astral_embed_enqueue_image(static_cast<AstralHandle>(EmbedderHandle), &Native, &Ticket);
     if (Err != ASTRAL_OK)
     {
@@ -163,7 +163,7 @@ bool UAstralEmbedder::EnqueueAudio(const FAstralAudioDesc& Audio, int64& OutTick
         return false;
     }
 
-    uint64 FrameCount = Audio.FrameCount;
+    uint64 FrameCount = static_cast<uint64>(Audio.FrameCount);
     if (FrameCount == 0)
     {
         const uint32 BytesPerSample = (Audio.Format == EAstralAudioFormat::F32) ? 4u : 2u;
@@ -175,18 +175,18 @@ bool UAstralEmbedder::EnqueueAudio(const FAstralAudioDesc& Audio, int64& OutTick
         FrameCount = TotalSamples / Audio.Channels;
     }
 
-    Native.channels = Audio.Channels;
-    Native.sample_rate = Audio.SampleRate;
+    Native.channels = static_cast<uint32_t>(Audio.Channels);
+    Native.sample_rate = static_cast<uint32_t>(Audio.SampleRate);
     Native.frame_count = FrameCount;
     Native.samples.data = Audio.Samples.GetData();
     Native.samples.len = static_cast<uint32_t>(Audio.Samples.Num());
-    Native.flags = Audio.Flags;
+    Native.flags = static_cast<uint32_t>(Audio.Flags);
     Native.gpu_device = Audio.GpuDevice;
-    Native.gpu_route_flags = Audio.GpuRouteFlags;
-    Native.gpu_device_mask = Audio.GpuDeviceMask;
+    Native.gpu_route_flags = static_cast<uint32_t>(Audio.GpuRouteFlags);
+    Native.gpu_device_mask = static_cast<uint64_t>(Audio.GpuDeviceMask);
     Native.gpu_stream = reinterpret_cast<void*>(static_cast<uintptr_t>(Audio.GpuStream));
 
-    uint64 Ticket = 0;
+    uint64_t Ticket = 0;
     const AstralErr Err = astral_embed_enqueue_audio(static_cast<AstralHandle>(EmbedderHandle), &Native, &Ticket);
     if (Err != ASTRAL_OK)
     {
@@ -227,15 +227,15 @@ bool UAstralEmbedder::EnqueueMultimodal(const FString& Text,
         }
         ImageNative.size = sizeof(AstralImageDesc);
         ImageNative.format = static_cast<AstralImageFormat>(Image.Format);
-        ImageNative.width = Image.Width;
-        ImageNative.height = Image.Height;
-        ImageNative.row_stride = Image.RowStride;
-        ImageNative.flags = Image.Flags;
+        ImageNative.width = static_cast<uint32_t>(Image.Width);
+        ImageNative.height = static_cast<uint32_t>(Image.Height);
+        ImageNative.row_stride = static_cast<uint32_t>(Image.RowStride);
+        ImageNative.flags = static_cast<uint32_t>(Image.Flags);
         ImageNative.pixels.data = Image.Pixels.GetData();
         ImageNative.pixels.len = static_cast<uint32_t>(Image.Pixels.Num());
         ImageNative.gpu_device = Image.GpuDevice;
-        ImageNative.gpu_route_flags = Image.GpuRouteFlags;
-        ImageNative.gpu_device_mask = Image.GpuDeviceMask;
+        ImageNative.gpu_route_flags = static_cast<uint32_t>(Image.GpuRouteFlags);
+        ImageNative.gpu_device_mask = static_cast<uint64_t>(Image.GpuDeviceMask);
         ImageNative.gpu_stream = reinterpret_cast<void*>(static_cast<uintptr_t>(Image.GpuStream));
         ImagePtr = &ImageNative;
     }
@@ -254,7 +254,7 @@ bool UAstralEmbedder::EnqueueMultimodal(const FString& Text,
         {
             return false;
         }
-        uint64 FrameCount = Audio.FrameCount;
+        uint64 FrameCount = static_cast<uint64>(Audio.FrameCount);
         if (FrameCount == 0)
         {
             const uint32 BytesPerSample = (Audio.Format == EAstralAudioFormat::F32) ? 4u : 2u;
@@ -265,20 +265,20 @@ bool UAstralEmbedder::EnqueueMultimodal(const FString& Text,
             }
             FrameCount = TotalSamples / Audio.Channels;
         }
-        AudioNative.channels = Audio.Channels;
-        AudioNative.sample_rate = Audio.SampleRate;
+        AudioNative.channels = static_cast<uint32_t>(Audio.Channels);
+        AudioNative.sample_rate = static_cast<uint32_t>(Audio.SampleRate);
         AudioNative.frame_count = FrameCount;
         AudioNative.samples.data = Audio.Samples.GetData();
         AudioNative.samples.len = static_cast<uint32_t>(Audio.Samples.Num());
-        AudioNative.flags = Audio.Flags;
+        AudioNative.flags = static_cast<uint32_t>(Audio.Flags);
         AudioNative.gpu_device = Audio.GpuDevice;
-        AudioNative.gpu_route_flags = Audio.GpuRouteFlags;
-        AudioNative.gpu_device_mask = Audio.GpuDeviceMask;
+        AudioNative.gpu_route_flags = static_cast<uint32_t>(Audio.GpuRouteFlags);
+        AudioNative.gpu_device_mask = static_cast<uint64_t>(Audio.GpuDeviceMask);
         AudioNative.gpu_stream = reinterpret_cast<void*>(static_cast<uintptr_t>(Audio.GpuStream));
         AudioPtr = &AudioNative;
     }
 
-    uint64 Ticket = 0;
+    uint64_t Ticket = 0;
     const AstralErr Err = astral_embed_enqueue_multimodal(
         static_cast<AstralHandle>(EmbedderHandle), TextSpan, ImagePtr, AudioPtr, &Ticket);
     if (Err != ASTRAL_OK)
@@ -306,7 +306,7 @@ bool UAstralEmbedder::Collect(int64 Ticket, TArray<float>& OutVector)
     Out.len = static_cast<uint32_t>(OutVector.Num() * sizeof(float));
 
     const AstralErr Err =
-        astral_embed_collect(static_cast<AstralHandle>(EmbedderHandle), static_cast<uint64>(Ticket), Out);
+        astral_embed_collect(static_cast<AstralHandle>(EmbedderHandle), static_cast<uint64_t>(Ticket), Out);
     return Err == ASTRAL_OK;
 }
